hopr_network_types

Module udp

Contains UDP socket-related helpers.

Structs

ConnectedUdpStream

Mimics TCP-like stream functionality on a UDP socket by restricting it to a single counterparty and implements [tokio::io::AsyncRead] and [tokio::io::AsyncWrite]. The instance is always constructed using a UdpStreamBuilder.

UdpStreamBuilder

Builder object for the ConnectedUdpStream.

Enums

ForeignDataMode

Defines what happens when data from another SocketAddr arrives into the ConnectedUdpStream (other than the one that is considered a counterparty for that instance).

UdpStreamParallelism

Determines how many parallel readers or writer sockets should be bound in ConnectedUdpStream.