hopr_network_types

Module utils

Various network IO-related utilities

Structs

AsyncReadStreamer

Converts a [AsyncRead] into [Stream] by reading at most S bytes in each call to [Stream::poll_next].

DuplexIO

Joins [futures::AsyncRead] and [futures::AsyncWrite] into a single object.

Functions

copy_duplex

This is a proper re-implementation of Tokio’s copy_bidirectional_with_sizes, which does not leave the stream in half-open-state when one side closes read or write side. Instead, if either side encounters and empty read (EOF indication), the write-side is closed as well and vice versa.